Extract code from DataElementsTable
